Tinkering with Bullet transforms.  Added PhysicsWorld dirty bits for timestep and gravity.

+++ avara-engine/src/Activity.cc
@@ -182,7 +182,7 @@ void Activity::update() override {
 		
 #warning move to saveGIFFrame()
 		float time = sceneTime();
-		static double previousSeconds = time;
+		static float previousSeconds = time;
 		float deltaSeconds = time - previousSeconds;
 		previousSeconds = time;
